requirements.txt
Pillow==8.1.2
pyaes==1.6.1
pyasn1==0.4.8
rsa==4.7.2
Telethon==1.21.1
